ScintillaConfig fixes for updated SDK target

+++ Scintilla/ScintillaConfig.hpp
@@ -8,11 +8,11 @@ public:
 	static Data cfg_string_to_data(wil::zstring_view str);
 	static Data get_default_data();
 	static Data preset_to_data(int id);
-	static int get_mode();
-	static int get_zoom();
+	static int64_t get_mode();
+	static int64_t get_zoom();
 	static std::string data_to_string(const Data& data);
-	static void set_mode(int mode);
-	static void set_zoom(int zoom);
+	static void set_mode(int64_t mode);
+	static void set_zoom(int64_t zoom);
 
 	void export_to_file(wil::zstring_view path);
 	void import_from_file(wil::zstring_view path);
